Mess around with how pages are structured in the codebase
I dislike Rehike's style as I find it hard to keep track of where which things are, so I decided to experiment with this style where all files pertaining to a single page are kept nearby in a single related folder.

The Rehike "models" folder remains, for now, for the sake of Rebug, which I have been hesitant to edit.