Make forgot password link in the header's signin dialog use the common formatted string type
This is a tiny bit wasteful, because this is guaranteed to always be a single run (it only needs an associated link), but it's the most convenient way to implement it.

I would have gone with FormattedStringRun, which is identical in prototype to the anonymous object we were using previously, but that is explicitly an internal type in the FormattedString implementation, so I didn't want to risk it. (A lot of the Rehike framework is poorly designed in hindsight)